+++ middlewares/index.js
@@ -1,5 +1,9 @@
 const validateBody = require("./validateBody");
+const isBodyEmpty = require("./isBodyEmpty");
+const isValidId = require("./isValidId");
 
 module.exports = {
   validateBody,
+  isBodyEmpty,
+  isValidId,
 };
